foodgram/requirements.txt
Django==2.2.19
pytz==2022.7.1
sqlparse==0.4.3

django-colorfield==0.8.0
Pillow==9.4.0

yapf==0.32.0
# flake8==6.0.0

djangorestframework==3.12.3
djoser==2.1.0

django-filter==2.1.0

python-dotenv==0.21.0
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6

django-debug-toolbar==3.2.4

django-cors-headers==3.10.0

# pyyaml 
# pip install django-cors-headers==3.10.0